Allow setting a 'typeName' inside a YAML query definition
Certain mbean objectnames do not have a 'name' key property used to uniquely
identify a particular group of metrics when using a wildcard query.

This adds support for an optional 'typeName' attribute to be defined in a query
section (defaulting to 'name', to replicate previous behaviour), the value of
which propagates to the outputWriters section in the JSON output.

Example YAML definition:
queries:
    - name: activemq
      obj : "org.apache.activemq:BrokerName=embedded,Type=Queue,Destination=*"
      typeName: "Destination"
      resultAlias: "activemq"
      attr:
        - "QueueSize"