Toward immutable data structures
Immutable data structures ensure thread safety and tend to simplify program flow. For example, there is no reason that an OutputWriter would modify a Result, so it should not be allowed to do it.

This commit is still work in progress and will be squashed before a PR is sent. Comments are welcomed.