All package cycles are now broken !
Moved a LOT of things around but very few code changes. Some of the classes were moved for purely technical reasons (break cycles) and it might make sense to rearrange them a bit more based on semantic reasons as well.

Maven enforcer rule added to ensure no cycles are created in the future.