Boolean values can be treated as number.
This fixes #99.

Conversion follows standard rules : true is 0, false is 1. It should be easy to add configuration to treat it differently. There is a new options (-b) to force treating boolean as booleans, but the default is to treat them as numbers. As far as I understand it, booleans are just ignored by all backends, so treating them as number by default is just as good (I have to admit I did not spend much time trying to understand the consequences of this change).

It might make sense to enable this option to be configurable for each output writer instead of each globally. Feedback is welcomed.

No tests yet (will come this weekend in the worst case).