Remvoed live check on Sockets.
This check is problematic as it waits 100[ms] for a timeout, which is a very, very long time once you have some traffic.

This PR might fix #123 andfix #190. Note the emphasis on the _might_. I've had similar problems with JMXTrans that have been fixed by removing this check, but I have not enough information on those issue to ensure this is the same problem.