Make JMXConnectionParams.hashCode() and equals() robust to array environment values.
This is a very ugly workaround which fixes #190. To make sure arrays with same parameters are treated as equals, we convert them to lists in the hashCode and equals methods. We could improve performance by caching the result (environment is immutable) but as environment is expected to be small, the gain is probably not worth the additional complexity.