Remove Ant build
Fixes #229, see https://github.com/jmxtrans/jmxtrans/pull/228#issuecomment-75340260 for additional discussion.

This also removes all the jar versioned with the project (in the /lib directory).

I took the opportunity to remove the javadoc that was versioned with the project. The Maven build publishes the Javadoc both as jar to Sonatype OSS (https://oss.sonatype.org/content/repositories/snapshots/org/jmxtrans/jmxtrans/247-SNAPSHOT/) and to the generated site (http://www.jmxtrans.org/jmxtrans/apidocs/index.html). Those are automatically updated on each commit to master, so most probably more trustworthy than what is in the project. The removal of Javadoc is not strictly related to the removal of Ant, but I think it make sense to do it in the same PR.