Allow Obj Domain to be used as the default key
Added a new boolean option named "useObjDomain" to BaseOutputWriter.

If set to true then generated keys will use the domain portion of the
ObjectName to form part of the output key rather than the ClassName.

For instance take a query which contains:
"obj" : "java.lang:type=Memory,*"

Currently it use the ClassName of the MBean to form part of the key
replacing any . characters with _ so in this case we will see keys
containing: sun_management_MemoryImpl

If useObjDomain is set to true then instead the key in this case would
contain java.lang (note the . is not being replaced by _)

If useObjDomain is set to false then the current behavior remains
unchanged.

By default this option will be false so unless users explicitly set it
to true they will not see any difference from the current behavior when
using JMXTrans.