Moved objDomain from Query to Result
This ensures that the correct domain from the ObjectName will be used at
query execution time. The output is now correct when useObjDomain is set
to true and a query obj that expands to multiple MBeans is used.

This would also make it easier to move away from KeyUtils to
NamingStrategy as NamingStrategy operates only over Result Objects.