Stopped tests from relying on incorrect behavior
Two branches, cleanup_allow_dotted and fix_booleanAsNumber were merged
into jmxtrans master. They were developed independently and fixed two
different issues. Unfortunately some of the tests depended on the
incorrect behavior fixed in the other branch and so when both were
merged a test failed. This fixes the issue. The code was correct, it was
the test that needed to change.