Modular GraphiteWriter
Following the introduction of the OutputWriterFactories, Graphite is the first OutputWriter to benefit from this.

Some more testing / validation is required before I am happy with this new OutputWriter, so it is at the moment published under a new name (GraphiteWriter2). Once I am happy, this new writer will probably replace the existing GraphiteWrite.

WIP