InfluxDbOutputWriter using OutputWriterFactory
This is an example of how I would go about it, feel free to modify it to your liking.

I introduce a clear separation between the construction of the OutputWriter and its runtime behaviour. This should help in simplifying tests. I also removed the dependency on BaseOutputWriter (which brings mostly nothing in this context).

I did not modify the unit tests, so they should not even compile at this point.