Preserve the order of typeNames in Query
As "key properties" of an `ObjectName` are an unordered set [0], the
order of typeNames was not respected. But the order becomes important
when constructing metric namse. For `ObjectName`s like

  org.apache.cassandra.db:type=ColumnFamilies,keyspace=foo,columnfamily=bar

and `typeNames` of `["keyspace", "columnfamily"]`, the unpredictable
order meant that the metrics would get reported as `foo_bar`, or
`bar_foo`. To avoid such issues, we now retain the order of typeNames.

Fixes #463

[0] http://docs.oracle.com/javase/7/docs/api/javax/management/ObjectName.html