Fix: jmxtrans does not exit on uncaught exceptions
If an exception causes a thread to die, jmxtrans may be in an
inconsistent state and should be terminated (at which point systemd, or
whatever is responsible for ensuring continuous running of jmxtrans, can
propertly restart it).

Register a global uncaught exception handler to log the error message to
the logger and exit with a non-zero status.

Fixes #685.