Preventing InstanceAlreadyExistsException on Reload
When a file was added/deleted/modified in the watchdir, the reload
would attempt to add the same mbeans to the JmxMBeanServer. This
would cause jmxtrans to hang and the watchdir functionality to be
useless.

Restructured reload to remove stale beans. Also, removes stale
server instances from the executor repositories so the repositories
don't grow unbounded when running with --use-separate-executors.