Query can specify how JMX Exceptions are handled.
Our environment runs a lot of Java applications in Linux HA cluster. So
application can run on any of cluster nodes. We are not able to maintain
JMXTrans config files to match applications running on the node and
JMXTrans configuration is same on every HA node. So many of applications
are not reachable for JMX and JMXTrans logs are full of exceptions.

Currently, when Exception happens during Query, full stack trace is
written to the log. This patch allows to specify attribute
jmxErrorHandling to a query with three ways how to handle Query
exception: dump, warn and ignore. Dump is current behaving.
Warn just prints warning about imposibility to collect the data (1 line).
Ignore just ignores any message.

Example:

{
  "servers" : [ {
    ...
    "queries" : [ {
      ...
      "jmxErrorHandling" : "warn",
      ...
    } ]
  } ]
}