JMXTrans can execute simple JMX operations
JMXTrans can execute JMX operation, if it is without parameters. Return
value is returned as string with .toString().

Example of configuration:

"queries" : [ {
  ...
  "operation" : [ "listSessions" ],
  ...

It can be combined with "attr" on same object. When no "attr" is used,
just operation is executed. When "attr" and "operation" is used, all
listed attrs and operations are queried.

Note: Before this change, when there was no "attr" in query all
attributes were queried. With the patch, behaving is slightly changed.
When no "operation" and no "attr" is used, all attributes are queried.
When "operation" is used, but no "attr" is used, just operations are
queried.

Note 2: Patch must be applied after "Query can specify how JMX
Exceptions are handled." PR#805 and "Extended list of variables
usable in configuration files with ${...}" PR#804.